\name{FluidgmR-package}
\alias{FluidgmR-package}
\alias{FluidgmR}
\docType{package}
\title{
FluidgmR
}
\description{
This package is a complete set of functions that organizes Fluidgm data and computes prediction of drug response based on an ensemble of machine learning techniques. It also accounts for missing observations and whenever feasible automatically imputes missing data using 100 random runs of the expectation maximization algorithm.

}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{FluidgmR}
\packageIndices{FluidgmR}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{Ujjal Kumar Mukherjee, Amit Kumar Mitra}

Maintainer: \packageMaintainer{Ujjal Kumar Mukherjee (ukm@illinois.edu)}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
