\name{SCATTome-package}
\alias{SCATTome-package}
\alias{SCATTome}
\docType{package}
\title{
SCATTome
}
\description{
This package is a complete set of functions that organizes Fluidigm targeted transcriptome data and computes prediction of drug response of single-cells based on an ensemble of machine learning techniques. It also accounts for missing observations and whenever feasible automatically imputes missing data using 100 random runs of the expectation maximization algorithm.

}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{SCATTome}
\packageIndices{SCATTome}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{Amit Kumar Mitra, Ujjal Kumar Mukherjee, Brian Van Ness}

Maintainer: \packageMaintainer{Ujjal Kumar Mukherjee (ukm@illinois.edu), bvnlab (mitra008@umn.edu)}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
